code/python/requirements.txt
# pip install -r requirements.txt 

numpy
opencv-python== 4.6.0.66
opencv-contrib-python ==4.6.0.66
Pillow

# Fast api
autopep8
fastapi
uvicorn[standard]
fastapi-utils
python-multipart
aiofiles
python-socketio
fastapi-socketio
websockets
